requirements.txt
Flask
Werkzeug
google-cloud-ndb
google-cloud-datastore
Jinja2
gunicorn
numpy